updated database code file

+++ src/database.cpp
@@ -10,9 +10,9 @@ void Database::write(std::vector<std::vector<std::string>> mainList)
     // check if file successfully opened
     if(db.is_open())
     {
-        for (unsigned int user_index=0; user_index < mainList[user_index].size(); user_index++)
+        for (size_t user_index=0; user_index < mainList[user_index].size(); user_index++)
         {
-            for (unsigned int list_index=0; list_index < mainList[user_index][list_index].size(); list_index++)
+            for (size_t list_index=0; list_index < mainList[user_index][list_index].size(); list_index++)
             {
                 db << mainList[user_index][list_index] << '\n';
             }
